converting functions removing unused ones too.
