+++ lib/constants/parameter-types.js
@@ -1,8 +1,5 @@
 'use strict';
 
-// Constants
-exports.parameterTypes = {
-  INPUT_PARAM   : 'in',
-  OUTPUT_PARAM  : 'out'
-};
-
+// Constants.
+exports.INPUT_PARAM = 'in';
+exports.OUTPUT_PARAM = 'out';
