+++ lib/dh-400jdbc.js
@@ -5,7 +5,7 @@ const _ = require('lodash');
 const EventEmitter = require('events').EventEmitter;
 const java = require('java');
 const async = require('async');
-const parameterTypes = require('./constants/parameter-types').parameterTypes;
+const parameterTypes = require('./constants/parameter-types');
 const util = require('util');
 
 // Driver Information.
@@ -99,72 +99,6 @@ function JDBCConn() {
 }
 
 util.inherits(JDBCConn, EventEmitter);
-
-/**
- * Builds the insert statement with placeholders using a key value parameter list.
- * @param tableName - The name of the table to insert into.
- * @param parameters - The key value parameter array.
- * @returns {string} - The sql string.
- */
-JDBCConn.prototype.buildInsertStatement  = function (tableName, parameters) {
-  let sql = `INSERT INTO ${tableName}`;
-
-  // check that there are parameters.
-  if (Object.keys(parameters).length) {
-    // build the parameter lists.
-    let columnNamesString = '';
-    let valuePlaceHolders = '';
-
-    // keep track of the current index.
-    let index = 0;
-
-    // loop over all the keys.
-    for (let key in parameters) {
-      columnNamesString += key;
-      valuePlaceHolders += '?';
-
-      // increment the index.
-      index++;
-
-      // if its not the last parameter, add a comma.
-      if (index !== Object.keys(parameters).length) {
-        columnNamesString += ', ';
-        valuePlaceHolders += ', ';
-      }
-    }
-
-    // connect it to the initial statement.
-    sql += ' (' + columnNamesString + ') VALUES (' + valuePlaceHolders + ')';
-  }
-
-  return sql;
-};
-
-/**
- * Converts a key value array or object into an array of just the values.
- * Note: use this in combination with the buildInsertStatement function to keep the values and columns in order.
- * @param obj
- * @returns {Array}
- */
-JDBCConn.prototype.objectToValueArray = (obj) => {
-  let results = [];
-
-  for (let key in obj) {
-    if (obj.hasOwnProperty(key)) {
-      results.push(obj[key]);
-    }
-  }
-
-  return results;
-};
-
-/**
- * Sets debug mode.
- * @param debug - The flag.
- */
-JDBCConn.prototype.setDebugMode = (debug) => {
-  this._debug = debug;
-};
 
 /**
  * Kills the jdbc connection.
