+++ lib/jdbc.js
@@ -4,8 +4,7 @@
 const _ = require('lodash');
 const java = require('java');
 const async = require('async');
-const parameterTypes = require('./constants/parameter-types').parameterTypes;
-const { EventEmitter } = require('events');
+const { INPUT_PARAM, OUTPUT_PARAM} = require('./constants/parameter-types');
 
 // Driver Information.
 const DRIVER_PATH             = `${__dirname}/../driver/jt400.jar`;
@@ -18,15 +17,12 @@ java.classpath.push(DRIVER_PATH);
 /**
  * JDBC object class.
  */
-class JDBC extends EventEmitter {
+class JDBC {
   /**
    * Constructor.
    * @param options - The config options.
    */
   constructor(options) {
-    // call the super function.
-    super();
-
     // default values.
     this.poolDataSource = null;
     this.pool = null;
@@ -72,16 +68,21 @@ class JDBC extends EventEmitter {
   }
 
   /**
-   * Kills the jdbc connection.
+   * Closes the sql connections.
    */
-  kill() {
-    // if there was not connection yet. skip.s
+  close() {
+    // if we aren't connected.
     if (!this.connected) {
-      this.logger.info('Not connected to AS400. Skipping kill.');
-      return;
+      this.logger.info('Not connected to the AS400. Skipping close.');
     }
 
-    // do something. TODO
+    try {
+      // close the connection pool.
+      this.pool.closeSync();
+    }
+    catch(ex) {
+      this.logger.error('Failed to close the connection pool.');
+    }
   }
 }
