+++ lib/jdbcUtilities.js
@@ -4,7 +4,7 @@
 const JDBCDriver = require('./dh-400jdbc');
 const jdbc = new JDBCDriver();
 const sqlTypes = require('./constants/sql-types').types;
-const parameterTypes = require('./constants/parameter-types').parameterTypes;
+const parameterTypes = require('./constants/parameter-types');
 
 // JDBC config.
 let config = {};
@@ -72,14 +72,6 @@ exports.setLogger = (loggerRef) => {
 };
 
 /**
- * Sets the debug mode.
- * @param debugMode - Debug mode on flag.
- */
-exports.setDebugMode = (debugMode) => {
-  jdbc.setDebugMode(debugMode);
-};
-
-/**
  * Sets the config object to the one specified.
  * @param configObject - The driver config object.
  */
@@ -89,10 +81,9 @@ exports.setConfig = (configObject) => {
 
 /**
  * Closes all connections in the pool.
- * @param callback
  */
-exports.closeAll = (callback) => {
-  jdbc.close(callback);
+exports.closeAll = () => {
+  jdbc.close();
 };
 
 /**
@@ -108,8 +99,8 @@ exports.insertData = (tableName, parameters, callback) => {
   }
 
   // build the statement and params.
-  let sql = jdbc.buildInsertStatement(tableName, parameters);
-  let values = jdbc.objectToValueArray(parameters);
+  let sql = buildInsertStatement(tableName, parameters);
+  let values = objectToValueArray(parameters);
 
   // execute the statement.
   exports.executeUpdatePreparedStatement(sql, values, (err) => {
@@ -315,3 +306,64 @@ exports.runTransaction = (executeFunction, callback) => {
     jdbc.runTransaction(executeFunction, callback);
   });
 };
+
+//======================================================================================
+// Helper Functions.
+//======================================================================================
+
+/**
+ * Returns an insert sql statement.
+ * @param tableName - The table name.
+ * @param parameters - The parameters object.
+ * @return {string}
+ */
+function buildInsertStatement(tableName, parameters) {
+  // start building the insert statement.
+  let sql = `INSERT INTO ${tableName}`;
+
+  // check that there are parameters.
+  if (Object.keys(parameters).length > 0) {
+    // build the parameter lists.
+    let columnNamesString = '';
+    let valuePlaceHolders = '';
+
+    // keep track of the current index.
+    let index = 0;
+
+    // loop over all the keys.
+    for (let key in parameters) {
+      columnNamesString += key;
+      valuePlaceHolders += '?';
+
+      // increment the index.
+      index++;
+
+      // if its not the last parameter, add a comma.
+      if (index !== Object.keys(parameters).length) {
+        columnNamesString += ', ';
+        valuePlaceHolders += ', ';
+      }
+    }
+
+    // connect it to the initial statement.
+    sql += ' (' + columnNamesString + ') VALUES (' + valuePlaceHolders + ')';
+  }
+
+  return sql;
+}
+
+/**
+ * Converts a JS object to value array.
+ * @param obj - The object.
+ */
+function objectToValueArray(obj) {
+  let results = [];
+
+  for (let key in obj) {
+    if (obj.hasOwnProperty(key)) {
+      results.push(obj[key]);
+    }
+  }
+
+  return results;
+}
